Refactor: Moved `auth_database.py` to `backend/data/` for better project organization
